move logger to the file
