+++ src/app.js
@@ -6,6 +6,7 @@ const userRouter = require('./resources/users/user.router');
 const boardRouter = require('./resources/boards/board.router');
 const columnRouter = require('./resources/columns/column.router');
 const taskRouter = require('./resources/tasks/task.router');
+const logger = require('./utils/logger');
 
 const app = express();
 const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
@@ -22,44 +23,7 @@ app.use('/', (req, res, next) => {
   next();
 });
 
-app.use('/boards/', (req, res, next) => {
-  next();
-});
-app.use((req, res, next) => {
-  const parseFunction = object => {
-    if (object === null) {
-      return null;
-    }
-    const returnString = Object.keys(object).reduce((acc, i) => {
-      let value = object[i];
-      if (typeof object[i] === 'object') {
-        value = parseFunction(object[i]);
-      }
-      return `${acc} ${Number.isInteger(parseInt(i, 10)) ? '' : i}${
-        Number.isInteger(parseInt(i, 10)) ? '' : ':'
-      }${value}`;
-    }, '');
-    const returnValue = Object.keys(object).length > 0 ? returnString : null;
-    return returnValue;
-  };
-  const nullOrValue = value => {
-    let returnValue = value;
-    if (value !== null && typeof value === 'object' && value !== undefined) {
-      returnValue = parseFunction(value);
-    }
-    return returnValue;
-  };
-  const url = req.originalUrl;
-  let { params, body, method } = req;
-  params = nullOrValue(params);
-  body = nullOrValue(body);
-  method = nullOrValue(method);
-  console.log(
-    `url:${url}; method:${method} query params:${params}; body:${body}`
-  );
-  next();
-});
-
+app.use(logger);
 app.use('/users', userRouter);
 app.use('/boards', boardRouter);
 app.use('/tasks', taskRouter);
@@ -73,7 +37,7 @@ app.use((err, req, res, next) => {
     res
   };
   if (err) {
-    process.stderr(err);
+    // process.stderr(err);
     res.status(500).send(value.message);
   }
   next();
